fix: rename files + cleaner code
