+++ prototype/src/components/Models/Maserati MC20.tsx
